pyproject.toml
[project]
name = "interview"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "black>=24.10.0",
    "langchain-core>=0.3.15",
    "langchain-anthropic>=0.2.4",
    "langchain-openai>=0.2.5",
    "langgraph>=0.2.43",
    "ipython>=8.29.0",
    "icecream>=2.1.3",
    "langchain>=0.3.9",
    "langchain-community>=0.3.9",
    "pytz>=2024.2",
    "langchain-google-genai>=2.0.8",
    "grpcio==1.60.1",
    "plyer>=2.1.0",
    "pyline-notify>=0.0.6",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/interview"]
